i#2626: AArch64 v8 decode: Add validating tests for GPR instructions
As part of #2626, exhaustive tests for the instructions that use GPR
registers have been generated. Some of these tests have revealed issues
with the current implementation of several of these instructions. This
patch is the first of several and contains all of the tests that validate
existing instructions as correct.

issues: #2626
Change-Id: I93444055a4eb028502d9baedeee8292f3e1c4a63