Add docs for Google Workload Traces
Adds a page describing and pointing at the Google data center workload
traces that are being made available in drmemtrace format.
The page deliberately uses a distinct top-level name so that
we can point at https://dynamorio.org/google_workload_traces.html.

Adds a pointer to the front page as this is considered relevant and
interesting to many users.