i#3348 sym conflicts: Rename set_cache_size to proc_set_cache_size (#5276)
Fixes a symbol conflict with a python package PyTables that is hit
when statically linking with certain toolchains where non-exported
symbol hiding is not available.  The fix is to rename set_cache_size()
to proc_set_cache_size().  This also matches the proc_ prefix for
related functions, so we do not resort to the d_r_ prefix.

Issue: #3348