i#4719 qemu: Replace WFI with WFE for QEMU interop (#4956)
QEMU crashes when executing the WFI AArchXX instruction.
DR uses WFI in its spinlock loops.
We replace WFI with WFE here, which is roughly a superset
of WFI and should work similarly for us while not
breaking QEMU.

This fixes over a dozen tests that previously failed under QEMU with
"unhandled CPU exception 0x10001".  Here we add 12 tests on AArch64
and 2 tests on AArch32 to the RUNS_ON_QEMU label.  More tests
could be added in AArch32 if a memcache assert were fixed.

Issue: #4719