i#5302: Fix a32 assert on rel-addr opnd disassembly
Adds correct ifdef ARM handling of REL_ADDR_kind operands when
disassembling, and in legacy loadtoconst code.

Adds a test to api.ir (confirmed it fails without this fix).

Fixes other latent issues in api.ir on arm which unfortunately does
not run on QEMU yet and so is not run automatically in today's CI.
Tested manually on an arm system.

Fixes #5302