i#4134 drbbdup: Add acquire semantics to case loading
Updates drbbdup_options_t.atomic_load_encoding to provide acquire
semantics, which is all that's generally needed for a load of a flag.

Implements the option on AArchXX (nothing to do on x86).

Adds a sanity test to one of the existing tests, ensuring the code
generated at least executes succesfully in one thread.

Issue: #4134