i#4139: Add method to clear droption values for static re-attach
The values of droption accumulated options are not reset on detach,
causing them to accumulate in a statically-linked client.

There is no control point for droption to do this automatically, so we
add droption_parser_t::clear_values() that a client can call.

Adds a call to clear_values() in drmemtrace.

Augments the drmemtrace burst_malloc test to fail when the options are
not reset (i.e., fail without this fix).  This serves as a test of the
new feature, as the existing droption tests do not have static clients
nor simple ways to add re-attaches.

Adds a section to the droption documentation explaining the new
feature.

Fixes #4139