i#1034: Eliminate signed char assumptions
Updates pieces of the code that used "char" and assumed it was signed
to use "sbyte" instead.

Adds "-funsigned-char" to the base compiler flags to really ensure the
code base is free of assumptions and to further ensure it is
interoperable with other toolchains that use "-funsigned-char".

Adds a note to the style guide.

Tweaks the drdecode_x86 test to fail without the fixes here in the x86
decoder.

Fixes #1034