i#5352 perf regression: Point at 9.0.1 release
We put out a 9.0.1 bugfix release for the #5352 performance regression
on Windows.  Here we update the direct download links.

Also updates the changelog, which includes a couple of minor features
added since 9.0.0.  We stick with 9.0.1 despite that for simplicity.
The only compatibility change affects just those building DR from
sources and so is less relevant for versioning.

Issue: #5352