i#5356 bbdup-wrap: Add extended drbbdup instrumentation callback
Adds a new field to the drbbdup options that points at an
instrumentation handler that takes in the "for_trace" and
"translating" parameters and returns dr_emit_flags_t.  Either the
original or this one must be set.

Adds a simple sanity check of the parameters.  Testing the return
value is difficult, unfortunately.

Issue: #5356, #3995, #4134