i#5356 bbdup-wrap: Add extended drbbdup instrumentation callbacks (#5361)
Adds new fields to the drbbdup options that point at a
case analysis handler and an instrumentation handler
that both take in the "for_trace" and "translating" parameters
and return dr_emit_flags_t.  For each, only one of the
original or the new extended version must be set.

Adds a simple sanity check of the parameters.  Testing the return
value is difficult, unfortunately.

Issue: #5356, #3995, #4134