i#4922: Fix tool.drcachesim.phys on AArch64.
Fixes the incorrect usage of OP_stur in tool.drcachesim.phys test on AArch64.
This was caused by delaying instrumentation for too many instructions, so that
when we finally did instrument them, we exceeded the  maximum immediate
displacement allowed by OP_stur. Our code needs to use OP_stur because the
trace_entry_t has a size of 12, meaning that every other entry in the buffer
will have a non-8-byte-aligned pc value.

Grants the online_instr_t class the ability to update the trace buffer pointer
by passing the required function to it.

Also removes the test from the ignore list for AArch64.

Fixes: #4922