i#5250: Add alt-bitwidth DR lib path config param
Adds a new config parameter DYNAMORIO_ALTINJECT for specifying the
path to the alternate-bitwidth DR library for cross-bitwidth execve.

If the parameter is set, DR uses it; otherwise it falls back to the
current hardcoded path assumptions.

drrun and friends automatically set the parameter if the other
bitwidth library is found.

Adds an automated test by augmenting the "linux.execve{32,64}" test
from a 64-bit build which does a test-and-build.  The renamed
"linux.xarch" test now does one build (for efficiency) and runs a
script which executes both (32<->64) directions for each of the
existing assumed-layout strategy and the new
config-points-at-alt-bitwidth strategy.

Also tested manually:
$ bin64/drrun -v -use_alt_dll ~/dr/releases/DynamoRIO-Linux-9.0.0/lib32/debug/libdynamorio.so -trace_threshold 51 -- suite/tests/bin/linux.execve64 ~/dr/git/build_x86_dbg_tests/suite/tests/bin/linux.execve-sub32

Fixes #5250