i#5398: Handle rep string expansion in drbbdup
Fixes a bug involving rep string expansion where drbbdup told drreg
to ignore control flow even when it was present in the app block.
Adds a guard on that setting based on a check for app control flow.

Adds a custom app for the drbbdup-emul-test which contains a zero-iter
rep string, and augments the test client to reserve rcx on the movs
only.  This reproduces the #5398 crash, and is fixed by the changes
here.

Fixes #5398