Add rstats and alloc type to beyond-vmm diagnostics
Adds diagnostics and fixes related to running out of memory with
drmemtrace now using drbbdup:

Augments the beyond-vmm diagnostics to dump the rstats and print the
allocation type.

Makes a beyond-vmm event non-fatal for -satisfy_w_xor_x for vmheap.

Resets a global drbbdup variable for reattach.

Removes a conditional for static_prepop that is not needed now that
stats are reset on reattach (#2964).

Issue: #2964