i#5421: Prevent app from closing dual_map_file.
Adds the dual_map_file to the DR fd_table. This is to prevent the app from
closing this FD, which manifests as an OOM on an extending commit.

Unfortunately, due to constraints on the order of invocation of
vmm_heap_unit_init and d_r_os_init in dynamorio_app_init_part_two_finalize, the
fd_table is not initialized when dual_map_file is created. This means that
fd_table_add will not really add it to the fd_table. So, we remember the
dual_map_file_fd and add it to fd_table when we create it.

Fixes: #5421