i#5195 win2019: Fix client.tls test.
Fixes the expected output for client.tls on Windows. The previous template
accounts for a mysterious output line, which we're not seeing anymore on
vs2019, so we just remove it from the expected output.

Also removes support for PRE_VS2015, seeing as we're effectively dropping
support for vs2017 by removing the above-mentioned line.

Issue: #5195