i#5463: Avoid globals pre-relocation
Fixes errors that can lead to crashes in relocate_dynamorio, where its
access to a function pointer and to the page size can both return
garbage pre-relocation.  We solve this with assembly to get the
current PC, and just using a 4K minimum page size for the backward
walk.

Tested with the PR #5462 branch where the new client.attach_blocking
test crashes without this fix on AArch64.
Manually tested on arm as well.

Fixes #5463