Update the victim block when current access is a hit
When the current access is a hit and the currently accessed 'way' has a
counter of 1, meaning it was the 'first accessed way' before this access,
set the next 'way' as victim. This is because the currently accessed
'way' is no longer the 'first accessed way' and its counter is then
cleared and the next 'way' is a victim.