i#5474: Eliminate stat glibc 2.33 dependence (#5476)
Eliminates a dependence on glibc 2.33 coming from the use of stat() in
drfrontendlib.  This is done by using the dr_stat_syscall() from drlibc,
which seems cleaner than using .symver and --wrap.  (Really, the Linux
toolchains should provide a simpler linker flag to solve this.)

Does not add a GA CI test by forcing the existing 2.17 check since we
want random local builds to ignore this: this is a one-time solution to
some of our own local uses cases.

Fixes #5474