i#5427 compress: Add lz4 drmemtrace raw compression (#5479)
Adds a new -raw_compress lz4 option.  lz4 out-performs even
snappy_nocrc, and it has knobs to support a range of i/o constraints.
However, it has no allocator parameterization, so it cannot be used
when drmemtrace is statically linked into the app.

Adds an lz4 reader and tests.

Updates the build docs and the GA CI installation to include
liblz4-dev.  Fixes the cross-compile jobs to also properly include
zlib and libsnappy by manually downloading the binaries.
Adds the libraries to Mac as well.

Issue: #5427