i#5427 compress; Turn -raw_compress lz4 on by default
For non-static drmemtrace, -raw_compress is now set to lz4 by default,
as that is nearly always faster than no compression or any other
compression scheme we've tried.

If lz4 is not enabled in the build, and again drmemtrace is not
static, snappy_nocrc is the default.

Otherwise, compression is disabled by default, just like today.

Inverts the raw-lz4 tests to become raw-none tests.

Updates the documentation.

Issue: #5427