i#4014 dr$sim phys: Make physaddr_t per-thread (#5508)
The physaddr_t class is not thread-safe and was previously used racily
in the drmemtrace code.  We fix that by creating a separate instance
per thread.  A test with multiple threads is added.

This does result in a per-thread file descriptor being opened, which
may not scale well: it will exhaust DR's private file-descriptor space
and could possibly hit rlimits.  Improving scaling is left as future work.

Issue: #4014