i#4014 dr$sim phys: Check privileges for pagemap
Since the kernel lets an unprivileged user read the pagemap file but
supplies 0 values for the physical pages (and 0 is a possible
legitimate value), we add an explicit check for privileges and fail up
front if physical addresses are requested but not available.  This is
a change in behavior where before execution would continue with a
warning.

Adds a test of missing privileges.

Fixes a sentinel issue where a 0 physical page was stored as 1 in the
table: but 1 is a possible valid number as well, so -1 is used
instead.

Issue: #4014