i#5527: Fix Android build broken by preset ANDROID_NDK_ROOT (#5532)
Fixes the problem of NDK setup failing due to ANDROID_NDK_ROOT pointing to the wrong path. The GitHub VM system preset ANDROID_NDK_ROOT to a location which does not include the necessary resources that the setup process needs. We set the ANDROID_NDK_ROOT to the right location before running the setup script.

Fixed: #5527