i#4014 dr$sim phys: Handle page-spanning accesses (#5536)
Adds handling of page-spanning instruction fetch and data accesses.
For offline traces these are heuristics as we do not know the precise
sizes at tracing time.  Since an extra physical page value is not
harmful, the heuristics emit a second page if there is any decent
chance that it is reached.

Adds a data page span test by switching the drcacheoff.phys test to
use allasm_repstr and adding such an access there.  An instruction
cross is more difficult to synthesize.

Issue: #4014