i#4014 dr$sim phys: Add p2v fail, prefetch, and funcs to view tool
The view tool was not printing the virtual address for
physical-to-virtual translations, nor the address for software
prefetches, which made it hard to diagnose physical address failures
in general and in particular when the culprit virtual address was from
a prefetch as it did not appear in the text output.  We add that here.

Since some prefetches have long names, we abbreviate them, but still
need to widen the name field to align them all.

While at it, we add the function marker values to the view tool as
well.

Example output:
       10: T462774 ifetch        7 byte(s) @ 0x00007f6cae908a54 non-branch
       11: T462774 read          8 byte(s) @ 0x00007f6cae90ca20 by PC 0x00007f6cae908a54
       12: T462774 ifetch        4 byte(s) @ 0x00007f6cae908a5c non-branch
       13: T462774 pref-w-L3-NT  8 byte(s) @ 0x00007f6c389xd238 by PC 0x00007f6cae908a5c
      ...
   441522: T48816 <marker: function #9>
   441523: T48816 <marker: function return address 0xaaaad5de4944>
   441524: T48816 <marker: function argument 0x15187bd4ae1d>
   441526: T48816 <marker: function argument 0x1>
      ...
   441546: T48816 <marker: function #9>
   441547: T48816 <marker: function return value 0x15187bd4ae1d>

Issue: #4014