i#5326: AArch64 codec: refactor codecsort.py to apply to all codecs
Previously codecsort.py was only applied to a single
codec at a time and thus couldn't allocate enums
uniquely as it didn't know about the enum values in
other files. This patch simplifies the process of
calling codecsort.py so that it can consider all the
codecs at once, and thus rewrite them as needed.

It also changes the default mode so that it returns
an exit code of 1 if it detects a rewrite is needed
making checking th order easier.

Issue: #5326

Change-Id: Id433cb9123e7cf901ae0e6bf14e84590cf39eadb