Add a comment warning on section alignment
Now that it's clear why sections are aligned to a 4K page-size, add a
comment warning that it might lead to issues on systems with a different
page size.

RISC-V is unaffected because so far only a 4K page is specified. Aarch64
is unaffected because page protections are only enabled on Windows so
far.

Signed-off-by: Stanislaw Kardach <kda@semihalf.com>