i#5570: Unhide static DR symbols.
Switches to dynamorio_static_unhide for configuring static DR so
that DR's symbols are visible when building static binaries.

Various symbols in dynamorio_static, like d_r_safe_read and
safe_read_if_fast in core/unix/os.c are non-weak symbols, but
they are not exported by the static DR library.
$ nm --defined ../../lib64/debug/libdynamorio_static.a | grep d_r_safe_read
00000000002962e8 t d_r_safe_read
$ nm --defined ../../lib64/debug/libdynamorio_static.a | grep safe_read_if_fast
0000000000296272 t safe_read_if_fast

This causes drlibc code to use the wrong routines in is_elf_so_header.
The same would happen for other weakly linked routines in drlibc which
are actually are supposed to be suppressed by their respective DR
definitions.

On Ubuntu 20, there's a non-readable vsyscall entry in maps. When drlibc
tries to read it, it crashes, and our main_signal_handler isn't able to
recognize it as a safe_read crash.

After this fix, the correct d_r_safe_read is used, which helps the DR
signal handler to recover as intended.

Fixes: #5570