i#5505 drpt2ir: Support mode without sideband information (#5562)
Update library drpt2ir to support setting the elf image for the whole PT trace.
And updates the command line client drpt2trace to support two conversion modes:
(1) Sideband Mode: the user must provide sideband data and parameters. In this mode, the converter uses sideband decoders to simulate image switches during the conversion. For example, we can use this mode to convert the traces where the instruction bytes are located in multiple images.
(2) Elf Mode: the user needs to provide an elf image for the PT trace. This mode is for cases where the PT trace's instruction bytes belong to one image. So we can use this mode to convert the kernel trace and the short-term user trace where it's likely that we'll not have an image switch.
Also, adds a new test case for the elf mode.

Issue: #5505