i#2062 memtrace nonmod: Add test for generated code
Adds a drmemtrace test which generates code outside of any module.

Adds a proper warning about this scenario at level 0 (with a do-once
to avoid spew) to both the tracer and raw2trace (previously it would
silently be traced without any indication of a problem, and the
raw2trace warning was at level 1).

Issue: #2062