i#2062 memtrace nonmod: Emit encodings for non-module code
Augments drmemtrace to emit the instruction encodings for non-module
code while tracing.  The instru_offline_t class emits them into a
global buffer and from there into a new file in the raw/ subdirectory.
Encodings are obtained by encoding each app or emulated instruction in
the block.

Adds a new drmemtrace interface function
drmemtrace_get_encoding_path() to obtain the path to the file where
the encodings are stored.  Adds a sanity check that the file exists to
the gencode test.

The file reader and raw2trace_directory_t are updated to be aware of
the encoding file.

Raw2trace support for parsing the encoding file will be added in a
separate commit to keep things small and incremental.

Issue: #2062