i#5620: Fix drmemtrace marker placement issues
Removes appending of non-kernel-xfer markers mid-block in raw2trace:
it was added as a mistake it seems.  We assume that all our markers
are at block boundaries, except kernel transfers which can happen
mid-block.

Adds three new invariant checks: that function markers are not between
an instruction and memref, that function markers are after branches,
and that function return addresses match the prior call instruction's
next pc.

Adds tests to raw2trace_unit_tests and to invariant_checker_tests.

Fixes #5620