i#2062 memtrace nonmod part 2: Raw2trace parsing of encoding file (#5619)
Adds raw2trace parsing of the encoding file used by the tracer to
store instruction encodings for generated code.  This involves the
following changes:

+ Adds encoding file parsing to module_mapper_t.
+ Changes module map queries to use new module_mapper_t interfaces instead,
  which handle generated code.
+ Changes block lookup to use the modidx,modoffs pair as the key rather than
  the absolute pc.  This runs into problems on 32-bit where the hashtable_t key
  is limited to pointer-sized.  To solve this, on 32-bit we use unordered_map, via
  a wrapper class block_hashtable_t to abstract away the differences.

The changes are compatibility-breaking for raw2trace_t which now takes
an encoding file parameter in the middle of existing parameters.
Updates existing uses.

For module_mapper_t the encoding file is added last with a default
value to preserve compatibility for existing analysis tools like
opcode_mix and view.  It is assumed that encodings for generated code
will be added to the final trace file and thus these tools will not
need a module_mapper_t interface for generated code.

Augments the tool.drcacheoff.gencode test to post-process the trace
and ensure the generated code PC is observed.

Fixes a -loglevel 4 signal dump_unmaksed() crash on detach i#5618 hit
in the gencode test; confirmed the test is crash-free at loglevel 4
with the fix.

Issue: #2062
Fixes #5618