i#5538 memtrace seek, part 1: Write in pieces to zipfiles
Splits post-processed offline drmemtrace files into chunks of a fixed
instruction count.  These chunks are combined inside one zipfile per
thread, maintaining the current file-per-thread invariant.

The minizip library, a contributed part of the zlib sources, is added
as a submodule and used to write and read the zipfile (via new
zipfile_ostream_t and zipfile_file_reader_t classes, respectively).
If the submodule is not present, we fall back to a gzipped single file
(if we have a system zlib).

Issue: #5538