i#5383 mac a64, part 2: Fix compiler warnings (#5632)
Fixes many unused-variable warnings when building the core and
extension libraries.  This enables a config without tests or docs to
build.  The drsyms libraries are missing so clients/ is not yet built.

Increases the clang-format version from 12 to 14 to match local versions,
as 12 has bugs in clang-format-diff that showed up in many places in
this PR.

Adds alignment of adjusted stack size options to page sizes in options_enable_code_api_dependences().

Issue: #5383