i#5636: Fix elided duplicate timestamps
PR #5633 added code to skip the duplicate timestamps at the top of
each chunk: but its logic assumed there would never be two legitimate
timestamps with identical values.  That does happen, in particular in
our online-drcachesim tests on Windows.  This resulted in the
invariant_checker not seeing some timestamp entries, causing its
exception for non-fetched instrs across thread switches to not apply
and resulting in invariant error reports.

We fix this by skipping the first timestamp in each chunk by
instruction count instead.  We'll want the insruction and chunk
counts for #5538 and I was about to add those fields in any case.

Fixes #5636